feat(system): 实现文件点击排名系统
- 新增 Engine 结构体，负责整体排名系统的逻辑- 实现数据恢复、点击记录、TopN 查询等功能
- 添加 Rdb 模块，用于持久化存储
- 新增 Wal模块，实现日志滚动和回放
- 修改 File 结构体，将 Id 和 Count 类型改为 uint64
- 移除 base_test.go 中的测试代码
- 新增 api 模块，定义 HTTP 接口
- 更新 config 模块，添加 Rdb 和 Wal 相关配置

+++ util/IdGenerator.go
@@ -32,7 +32,7 @@ func GetIdGenerator() *IDGenerator {
 }
 
 // GenerateID generateID 生成唯一ID
-func (g *IDGenerator) GenerateID() int {
+func (g *IDGenerator) GenerateID() uint64 {
 
 	// 获取当前时间戳（毫秒）
 	timestamp := time.Now().UnixMilli()
@@ -55,6 +55,6 @@ func (g *IDGenerator) GenerateID() int {
 	g.lastTimestamp = timestamp
 
 	// 将时间戳左移，并与序列号拼接生成最终的ID（确保不会超出int范围）
-	id := int((timestamp-epoch)<<sequenceBits) | int(g.sequence)
+	id := uint64((timestamp-epoch)<<sequenceBits) | uint64(g.sequence)
 	return id
 }
